Prihlasenie, napojenie na MySQL

+++ src/main/java/sk/euba/fhi/model/interf/BankUcetData.java
@@ -6,11 +6,11 @@ import java.util.List;
 
 public interface BankUcetData {
     List<BankUcet> vsetky();
-    List<BankUcet> preFirmu(long id_firma);
-    BankUcet getBankUcet(long id);
-    List<String> nazvyUctov(long id_firma);
-    long idUctu(long id_firma, String ucet_nazov);
+    List<BankUcet> preFirmu(int id_firma);
+    BankUcet getBankUcet(int id);
+    List<String> nazvyUctov(int id_firma);
+    int idUctu(int id_firma, String ucet_nazov);
     void vloz(BankUcet p);
     void zmen(BankUcet p);
-    void zmaz(long id);
+    void zmaz(int id);
 }
