+++ src/main/java/sk/euba/fhi/model/interf/BankaData.java
@@ -5,9 +5,9 @@ import sk.euba.fhi.model.obj.Banka;
 import java.util.List;
 
 public interface BankaData {
-    List<Banka> vsetky(long id_firmy, long id_uctu, int rok);
-    Banka getBanka(long id);
+    List<Banka> vsetky(int id_firmy, int id_uctu, int rok);
+    Banka getBanka(int id);
     void vloz(Banka p);
     void zmen(Banka p);
-    void zmaz(long id);
+    void zmaz(int id);
 }
