+++ src/main/java/sk/euba/fhi/model/interf/DFData.java
@@ -5,9 +5,9 @@ import sk.euba.fhi.model.obj.DF;
 import java.util.List;
 
 public interface DFData {
-    List<DF> vsetky(long id_firmy, int rok);
-    DF getDF(long id);
+    List<DF> vsetky(int id_firmy, int rok);
+    DF getDF(int id);
     void vloz(DF p);
     void zmen(DF p);
-    void zmaz(long id);
+    void zmaz(int id);
 }
