+++ src/main/java/sk/euba/fhi/model/interf/FirmaData.java
@@ -5,11 +5,11 @@ import sk.euba.fhi.model.obj.Firma;
 import java.util.List;
 
 public interface FirmaData {
-    List<Firma> vsetky(long pouzivatel_id);
-    List<String> nazvyFiriem(long pouzivatel_id);
-    long idFirmy(long pouzivatel_id, String firma_nazov);
-    Firma getFirma(long id);
+    List<Firma> vsetky(int pouzivatel_id);
+    List<String> nazvyFiriem(int pouzivatel_id);
+    int idFirmy(int pouzivatel_id, String firma_nazov);
+    Firma getFirma(int id);
     void vloz(Firma p);
     void zmen(Firma p);
-    void zmaz(long id);
+    void zmaz(int id);
 }
