+++ src/main/java/sk/euba/fhi/model/interf/PokladnaData.java
@@ -5,9 +5,9 @@ import sk.euba.fhi.model.obj.Pokladna;
 import java.util.List;
 
 public interface PokladnaData {
-    List<Pokladna> vsetky(long id_firmy, int rok);
-    Pokladna getPokladna(long id);
+    List<Pokladna> vsetky(int id_firmy, int rok);
+    Pokladna getPokladna(int id);
     void vloz(Pokladna p);
     void zmen(Pokladna p);
-    void zmaz(long id);
+    void zmaz(int id);
 }
