+++ src/main/java/sk/euba/fhi/model/interf/VFData.java
@@ -5,9 +5,9 @@ import sk.euba.fhi.model.obj.VF;
 import java.util.List;
 
 public interface VFData {
-    List<VF> vsetky(long id_firmy, int rok);
-    VF getVF(long id);
+    List<VF> vsetky(int id_firmy, int rok);
+    VF getVF(int id);
     void vloz(VF p);
     void zmen(VF p);
-    void zmaz(long id);
+    void zmaz(int id);
 }
